Rename mrp_PD algorithms for clarity
