+++ src/fswAlgorithms/attControl/mrpPD/mrpPD.cpp
@@ -1,7 +1,7 @@
 /*
  ISC License
 
- Copyright (c) 2016, Autonomous Vehicle Systems Lab, University of Colorado at Boulder
+ Copyright (c) 2024, Laboratory for Atmospheric and Space Physics, University of Colorado at Boulder
 
  Permission to use, copy, modify, and/or distribute this software for any
  purpose with or without fee is hereby granted, provided that the above
@@ -16,84 +16,110 @@
  OR IN CONNECTION WITH THE USE OR PERFORMANCE OF THIS SOFTWARE.
 
  */
-/*
-    mrpPD Module
 
- */
-
-#include "fswAlgorithms/attControl/mrpPD/mrpPD.h"
+#include "mrpPD.h"
+#include "architecture/utilities/avsEigenSupport.h"
 #include "architecture/utilities/linearAlgebra.h"
-#include "architecture/utilities/macroDefinitions.h"
-#include <string.h>
 
 /*! This method performs a complete reset of the module.  Local module variables that retain
  time varying states between function calls are reset to their default values.
  @return void
  @param callTime [ns] Time the method is called
 */
-void MrpPD::reset(uint64_t callTime)
-{
-    // check if the required input messages are included
+void MrpPD::reset(uint64_t callTime) {
+    // Check if the required input messages are linked
     if (!this->guidInMsg.isLinked()) {
-        this->bskLogger.bskLog(BSK_ERROR, "Error: mrpPD.guidInMsg wasn't connected.");
+        _bskLog(this->bskLogger, BSK_ERROR, "mrpPD.guidInMsg wasn't connected.");
     }
-
     if (!this->vehConfigInMsg.isLinked()) {
-        this->bskLogger.bskLog(BSK_ERROR, "Error: mrpPD.vehConfigInMsg wasn't connected.");
+        _bskLog(this->bskLogger, BSK_ERROR, "mrpPD.vehConfigInMsg wasn't connected.");
     }
 
-
-    /*! - read in spacecraft configuration message */
-    VehicleConfigMsgPayload vcInMsg = this->vehConfigInMsg();
-    mCopy(vcInMsg.ISCPntB_B, 1, 9, this->ISCPntB_B);
+    // Read the VehicleConfigMsgPayload input message
+    if (this->vehConfigInMsg.isWritten()) {
+        VehicleConfigMsgPayload vcInMsg;
+        vcInMsg = this->vehConfigInMsg();
+        this->ISCPntB_B = cArray2EigenMatrixXd(vcInMsg.ISCPntB_B, 3, 3);
+    }
 }
 
-/*! This method takes the attitude and rate errors relative to the Reference frame, as well as
-    the reference frame angular rates and acceleration, and computes the required control torque Lr.
+/*! This method takes the attitude and rate errors relative to the reference frame, as well as
+the reference frame angular rates and acceleration, and computes the required control torque Lr.
  @return void
- @param callTime The clock time at which the function was called (nanoseconds)
+ @param callTime [ns] Time the method is called
 */
-void MrpPD::updateState(uint64_t callTime)
-{
-    double              Lr[3];                  /*!< required control torque vector [Nm] */
-    double              omega_BN_B[3];          /*!< Inertial angular body vector in body B-frame components */
-    CmdTorqueBodyMsgPayload controlOutMsg = {}; /*!< Control output requests */
-    AttGuidMsgPayload       guidInMsg;          /*!< Guidance Message */
-    double              v3_temp1[3];
-    double              v3_temp2[3];
-    double              v3_temp3[3];
-    double              v3_temp4[3];
+void MrpPD::updateState(uint64_t callTime) {
+    // Create the buffer messages
+    CmdTorqueBodyMsgPayload torqueCmdMsgPayload;  // Control output request msg
+    AttGuidMsgPayload guidanceMsgPayload;         // Guidance input message
 
-    /*! - Read the guidance input message */
-    guidInMsg = this->guidInMsg();
+    // Zero the output message
+    torqueCmdMsgPayload = CmdTorqueBodyMsgPayload();
 
-    /*! - Compute angular body rate */
-    v3Add(guidInMsg.omega_BR_B, guidInMsg.omega_RN_B, omega_BN_B);
+    // Read the guidance input message
+    guidanceMsgPayload = AttGuidMsgPayload();
+    if (this->guidInMsg.isWritten()) {
+        guidanceMsgPayload = this->guidInMsg();
+    }
 
-    /*! - Evaluate required attitude control torque */
-    /* Lr =  K*sigma_BR + P*delta_omega  - omega_r x [I]omega - [I](d(omega_r)/dt - omega x omega_r) + L
-     */
-    v3Scale(this->K, guidInMsg.sigma_BR, v3_temp1); /* + K * sigma_BR */
-    v3Scale(this->P, guidInMsg.omega_BR_B, v3_temp2); /* + P * delta_omega */
-    v3Add(v3_temp1, v3_temp2, Lr);
+    // Compute hub inertial angular velocity in B-frame components
+    Eigen::Vector3d omega_BR_B = cArray2EigenVector3d(guidanceMsgPayload.omega_BR_B);
+    Eigen::Vector3d omega_RN_B = cArray2EigenVector3d(guidanceMsgPayload.omega_RN_B);
+    Eigen::Vector3d omega_BN_B = omega_BR_B + omega_RN_B;
 
-    /* omega x [I]omega */
-    m33MultV3(RECAST3X3 this->ISCPntB_B, omega_BN_B, v3_temp3);
-    v3Cross(guidInMsg.omega_RN_B, v3_temp3, v3_temp3); /* omega_r x [I]omega */
-    v3Subtract(Lr, v3_temp3, Lr);
+    // Compute K*sigma_BR
+    Eigen::Vector3d sigma_BR = cArray2EigenVector3d(guidanceMsgPayload.sigma_BR);
+    Eigen::Vector3d v3_temp1 = this->K * sigma_BR;
 
-    /* [I](d(omega_r)/dt - omega x omega_r) */
-    v3Cross(omega_BN_B, guidInMsg.omega_RN_B, v3_temp4);
-    v3Subtract(guidInMsg.domega_RN_B, v3_temp4, v3_temp4);
-    m33MultV3(RECAST3X3 this->ISCPntB_B, v3_temp4, v3_temp4);
-    v3Subtract(Lr, v3_temp4, Lr);
+    // Compute P*delta_omega
+    Eigen::Vector3d v3_temp2 = this->P * omega_BR_B;
 
-    v3Add(this->knownTorquePntB_B, Lr, Lr); /* + L */
-    v3Scale(-1.0, Lr, Lr);
+    // Compute omega_r x [I]omega
+    Eigen::Vector3d v3_temp3 = omega_RN_B.cross(this->ISCPntB_B * omega_BN_B);
 
-    /*! - Store and write the output message */
-    v3Copy(Lr, controlOutMsg.torqueRequestBody);
-    this->cmdTorqueOutMsg.write(&controlOutMsg, moduleID, callTime);
+    // Compute [I](d(omega_r)/dt - omega x omega_r)
+    Eigen::Vector3d domega_RN_B = cArray2EigenVector3d(guidanceMsgPayload.domega_RN_B);
+    Eigen::Vector3d v3_temp4 = this->ISCPntB_B * (domega_RN_B - omega_BN_B.cross(omega_RN_B));
 
-    return;
+    // Compute required attitude control torque vector
+    // Lr =  K*sigma_BR + P*delta_omega  - omega_r x [I]omega - [I](d(omega_r)/dt - omega x omega_r) + L
+    Eigen::Vector3d Lr =
+        -v3_temp1 - v3_temp2 + v3_temp3 + v3_temp4 - this->knownTorquePntB_B;  // [Nm] Required control torque vector
+
+    // Write the output message
+    eigenVector3d2CArray(Lr, torqueCmdMsgPayload.torqueRequestBody);
+    this->cmdTorqueOutMsg.write(&torqueCmdMsgPayload, moduleID, callTime);
 }
+
+/*! Getter method for the derivative gain P.
+ @return const double
+*/
+double MrpPD::getDerivativeGainP() { return this->P; }
+
+/*! Getter method for the known torque about point B.
+ @return const Eigen::Vector3d
+*/
+const Eigen::Vector3d &MrpPD::getKnownTorquePntB_B() const { return this->knownTorquePntB_B; }
+
+/*! Getter method for the proportional gain K.
+ @return const double
+*/
+double MrpPD::getProportionalGainK() { return this->K; }
+
+/*! Setter method for the derivative gain P.
+ @return void
+ @param P [N*m*s] Rate error feedback gain applied
+*/
+void MrpPD::setDerivativeGainP(double P) { this->P = P; }
+
+/*! Setter method for the known external torque about point B.
+ @return void
+ @param knownTorquePntB_B [N*m] Known external torque expressed in body frame components
+*/
+void MrpPD::setKnownTorquePntB_B(Eigen::Vector3d &knownTorquePntB_B) { this->knownTorquePntB_B = knownTorquePntB_B; }
+
+/*! Setter method for the proportional gain K.
+ @return void
+ @param K [rad/s] Proportional gain applied to MRP errors
+*/
+void MrpPD::setProportionalGainK(double K) { this->K = K; }
