+++ src/fswAlgorithms/attControl/mrpPD_C/mrpPD_C.h
@@ -1,0 +1,52 @@
+/*
+ ISC License
+
+ Copyright (c) 2016, Autonomous Vehicle Systems Lab, University of Colorado at Boulder
+
+ Permission to use, copy, modify, and/or distribute this software for any
+ purpose with or without fee is hereby granted, provided that the above
+ copyright notice and this permission notice appear in all copies.
+
+ THE SOFTWARE IS PROVIDED "AS IS" AND THE AUTHOR DISCLAIMS ALL WARRANTIES
+ WITH REGARD TO THIS SOFTWARE INCLUDING ALL IMPLIED WARRANTIES OF
+ MERCHANTABILITY AND FITNESS. IN NO EVENT SHALL THE AUTHOR BE LIABLE FOR
+ ANY SPECIAL, DIRECT, INDIRECT, OR CONSEQUENTIAL DAMAGES OR ANY DAMAGES
+ WHATSOEVER RESULTING FROM LOSS OF USE, DATA OR PROFITS, WHETHER IN AN
+ ACTION OF CONTRACT, NEGLIGENCE OR OTHER TORTIOUS ACTION, ARISING OUT OF
+ OR IN CONNECTION WITH THE USE OR PERFORMANCE OF THIS SOFTWARE.
+
+ */
+
+#ifndef _MRP_PD_CONTROL_C_H_
+#define _MRP_PD_CONTROL_C_H_
+
+#include "architecture/_GeneralModuleFiles/sys_model.h"
+#include "architecture/messaging/messaging.h"
+#include "architecture/msgPayloadDefC/AttGuidMsgPayload.h"
+#include "architecture/msgPayloadDefC/CmdTorqueBodyMsgPayload.h"
+#include "architecture/msgPayloadDefC/VehicleConfigMsgPayload.h"
+#include <stdint.h>
+
+/*! @brief Module configuration message definition. */
+class MrpPD_C : public SysModel {
+   public:
+    void reset(uint64_t callTime) override;
+    void updateState(uint64_t callTime) override;
+
+    /* declare public module variables */
+    double K;                     //!< [rad/sec] Proportional gain applied to MRP errors
+    double P;                     //!< [N*m*s]   Rate error feedback gain applied
+    double knownTorquePntB_B[3];  //!< [N*m]     known external torque in body frame vector components
+
+    /* declare private module variables */
+    double ISCPntB_B[9];  //!< [kg m^2] Spacecraft Inertia
+
+    /* declare module IO interfaces */
+    Message<CmdTorqueBodyMsgPayload> cmdTorqueOutMsg;     //!< commanded torque output message
+    ReadFunctor<AttGuidMsgPayload> guidInMsg;             //!< attitude guidance input message
+    ReadFunctor<VehicleConfigMsgPayload> vehConfigInMsg;  //!< vehicle configuration input message
+
+    BSKLogger bskLogger = {};  //!< BSK Logging
+};
+
+#endif
